Add prebuild dependency in order to cache binaries off to GitHub.
Rationale: The prebuild module is a super easy way to squirrel away a binary for future use. Users often don't have the build chains in order to compile the bindings. It would be splendid if people with write access to your project would do this. Of course, prebuild would default to the same system you have right now if no binaries exist. You can override as well. Either way, this would help!